[CHG] More UI improvements, refactorings
I removed the Search class, as it was just passing throught the
results from Minimax, so just call Minimax directly.  Search used to
be a way to have different search algorithms, but that is no longer
the case.

There are a few UI improvements to show if there is a winner
(unstyled).  Also the button is now disabled on a winner.  Again,
more UI improvements to come, but the game is much more functional now.